Hello
Hello

My name is Lisa Cornell, and I live in Wake Forest, North Carolina. I am a proud mother of five and a grandmother of three.  I work full-time and enjoy staying active with VR workouts, as well as reading and spending quality time with my children and grandchildren.

Academic integrity is a crucial value in education because it fosters trust, fairness, and respect within the academic community. For professors, it ensures credibility of their teaching and research.  For students, upholding academic integrity is essential for developing genuine knowledge and skills, building character and preparing for future professional responsibilities.